+++ js/data.js
@@ -1,0 +1,16 @@
+const TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+const CHECKINS = ['12:00', '13:00', '14:00'];
+const CHECKOUTS = ['12:00', '13:00', '14:00'];
+const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+const TITLE = 'Вам может быть интересен этот вариант';
+const DESCRIPTION = 'Лучший вариант в этом регионе';
+const MAX_PRICE = 10000000;
+const MAX_GUESTS = 25;
+const MAX_ROOMS = 10;
+
+export {TYPES, CHECKINS, CHECKOUTS, FEATURES, PHOTOS, TITLE, DESCRIPTION, MAX_PRICE, MAX_GUESTS, MAX_ROOMS};
